api/types/registry: move `SearchOptions` to client
